ai-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
numpy==1.26.3
python-dotenv==1.0.0
